feat: guards to post/patch/delete

+++ src/menu/menu.service.ts
@@ -1,10 +1,15 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Inject, Injectable } from '@nestjs/common';
+import { REQUEST } from '@nestjs/core';
+import { Request } from 'express';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateMenuDto } from './dto/create-menu.dto';
 
 @Injectable()
 export class MenuService {
-  constructor(private prisma: PrismaService) {}
+  constructor(
+    private prisma: PrismaService,
+    @Inject(REQUEST) private readonly request: Request,
+  ) {}
 
   async create(createMenuDto: CreateMenuDto) {
     return await this.prisma.menu.create({ data: createMenuDto });
@@ -19,6 +24,13 @@ export class MenuService {
   }
 
   async remove(id: number) {
+    const userId = this.request['user']['sub'];
+    const menuUserId = (
+      await this.prisma.menu.findUnique({
+        where: { id: id },
+      })
+    ).userId;
+    if (userId !== menuUserId) throw new ForbiddenException();
     return await this.prisma.menu.delete({ where: { id: id } });
   }
 }
